requirements.txt
astroid==2.1.0
autopep8==1.4.3
charset-normalizer==3.3.2
colorama==0.4.1
distlib==0.3.8
Django==2.2.28
djangorestframework==3.11.2
et-xmlfile==1.1.0
filelock==3.13.1
idna==3.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==5.1.0
mccabe==0.6.1
oauthlib==3.2.2
openpyxl==3.1.2
pep8==1.7.1
platformdirs==4.1.0
platformshconfig==2.4.0
pyasn1==0.5.1
pycodestyle==2.4.0
pylint==2.2.2
pyparsing==3.1.1
pytz==2018.7
six==1.12.0
sqlparse==0.4.4
StrEnum==0.4.15
urllib3==2.1.0
virtualenv==20.25.0
wrapt==1.10.11
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
python-dotenv==1.0.1
Werkzeug==3.0.4
